feat(seahaven-cli): introduce `eject` command
- Introduced a new `eject` command that allows users to extract the `setup.yaml` file into `docker-compose.yaml` and `.env` files.
- Implemented user confirmation for the one-way operation, ensuring users are aware of the implications of ejecting the setup.
- Updated command structure to include the new `eject` command in the CLI.
