+++ seahaven-cli/src/cmd/root.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use super::{build, down, init, pull, run, setup, up, version};
+use super::{build, down, eject, init, pull, run, setup, up, version};
 use crate::result::Result;
 
 /// Create and execute the DIPs CLI command line interface
@@ -9,6 +9,7 @@ pub async fn cmd_run() -> Result<()> {
         .subcommands([
             build::cmd(),
             down::cmd(),
+            eject::cmd(),
             init::cmd(),
             pull::cmd(),
             run::cmd(),
@@ -29,6 +30,7 @@ pub async fn cmd_run() -> Result<()> {
     match matches.subcommand() {
         Some((build::CMD, matches)) => build::run(matches).await,
         Some((down::CMD, matches)) => down::run(matches).await,
+        Some((eject::CMD, matches)) => eject::run(matches).await,
         Some((init::CMD, matches)) => init::run(matches).await,
         Some((pull::CMD, matches)) => pull::run(matches).await,
         Some((run::CMD, matches)) => run::run(matches).await,
